Remove AbstractApp and Refactor App
Consolidate some functionality into App. Move some functionality into mixins and change App's inheritance.
